feat: finaliza os DAO e cria os BO com as regras de negócio

+++ src/main/java/br/com/gwfrete/util/FreteUtil.java
@@ -3,7 +3,7 @@ package br.com.gwfrete.util;
 import java.time.LocalDateTime;
 
 public class FreteUtil {
-    public static String formatarNumeroFrete(int sequencial) {
+    public static String formatarNumeroFrete(long sequencial) {
 
         int ano = LocalDateTime.now().getYear();
         return String.format("FRT-%d-%05d", ano, sequencial);
